Extract JDM parsers for Node and Python
